requirements.txt
openai>=0.28.0
python-docx>=0.8.11
python-dotenv>=1.0.0
tqdm>=4.65.0